% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map.R
\name{base_map}
\alias{base_map}
\title{Base Map}
\usage{
base_map(bbox, increase_zoom = 0, basemap = "dark", nolabels = F)
}
\arguments{
\item{bbox}{bounding box for out map extents generated using st_bbox() and expanded as necessary using expand_bbox(). Bounding box should be in lat/lng (epsg: 4326).}

\item{increase_zoom}{the zoom is automatically calculated for the map but we can increase or decrease the zoom by setting an integer value.}

\item{basemap}{the style of basemap to use. Currently supports 'dark', 'hydda', 'positron', 'voyager', 'wikimedia', 'mapnik', 'neighbourhood'}

\item{nolabels}{if TRUE, removes labels from the basemap. This is only available for some styles.}
}
\value{
a set of tiles to be added to a ggplot object.
}
\description{
fetches a basemap for our map based on the map's bounding box. Includes themes available at: https://leaflet-extras.github.io/leaflet-providers/preview/
}
\examples{
# get bounding box for our map
bbox <- st_bbox(localauth_data)

# add to ggplot
ggplot() +
base_map(bbox, increase_zoom = 2, basemap = 'hydda')+
geom_sf(data = localauth_data, fill =NA )  +
coord_sf(xlim = c(bbox$xmin, bbox$xmax), ylim = c(bbox$ymin, bbox$ymax), crs = 4326)


# add straight to ggplot
ggplot() +
base_map(st_bbox(localauth_data), increase_zoom = 2) +
geom_sf(data = localauth_data, fill = NA)

}
