% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_bbox.R
\name{expand_bbox}
\alias{expand_bbox}
\title{Expand Bounding Box}
\usage{
expand_bbox(bbox, metresX, metresX2 = metresX, metresY,
  metresY2 = metresY, crs_out = 4326)
}
\arguments{
\item{bbox}{a bounding box generated by sf::st_bbox() or named numeric vector with xmin, ymin, xmax, ymax.}

\item{metresX}{the distance in metres that we want to expand the bounding box by in the X direction}

\item{metresX2}{if specified, the meters in the Easterly direction and metresX becomes metres in the Westerly direction.}

\item{metresY}{the distance in metres that we want to expand the bounding box by in the Y direction}

\item{metresY2}{if specified, the metres to the South. metresY becomes metres to the North.}

\item{crs_out}{what epsg coordinate system to return the bbox in. Defaults to lat/lng (4326).}
}
\value{
a named numeric vector of bbox extents. This can be converted into an sf object using sf::st_as_sfc(). Coordinates are in lat/lng (4326).
}
\description{
A function to take a bounding box (generated using st_bbox) and expand it by x meters in the X direction and y meters in the Y direction.
}
\examples{

bbox <- expand_bbox(st_bbox(), 5000, 5000)
dbDisconnect(con)

}
