% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{localauth_data}
\alias{localauth_data}
\title{London Local Authority GIS Data}
\format{A sf object with 5 variables:
\describe{
\item{Name}{Local Authority Name}
\item{Code}{Official Local Authority ID}
\item{Full_Description}{Concatenated Name and Code}
\item{geometry}{geometry column}
}}
\source{
\url{https://geoportal.statistics.gov.uk/datasets/local-authority-districts-april-2019-boundaries-uk-bfc}
}
\usage{
localauth_data
}
\description{
A dataset containing 2019 Local Authority boundaries for London
}
\keyword{datasets}
