% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_bbox.R
\name{expand_bbox}
\alias{expand_bbox}
\title{Expand bbox}
\usage{
expand_bbox(bbox, X, Y, X2 = X, Y2 = Y, crs_out = 4326)
}
\arguments{
\item{bbox}{a bounding box generated by sf::st_bbox() or named numeric vector with xmin, ymin, xmax, ymax.}

\item{X}{the distance in metres that we want to expand the bounding box by in the X direction}

\item{Y}{the distance in metres that we want to expand the bounding box by in the Y direction}

\item{X2}{if specified, the meters in the Easterly direction and metresX becomes metres in the Westerly direction.}

\item{Y2}{if specified, the metres to the South. metresY becomes metres to the North.}

\item{crs_out}{what epsg coordinate system to return the bbox in. Defaults to lat/lng (4326).}
}
\value{
a named numeric vector of bbox extents. This can be converted into an sf object using sf::st_as_sfc(). Coordinates are in lat/lng (4326).
}
\description{
A function to take a bounding box (generated using st_bbox) and expand it by x meters in the X direction and y meters in the Y direction.
}
\examples{
camden <- dplyr::filter(localauth_data, Name == 'Camden') \%>\%
  st_transform(4326)
bbox <- expand_bbox(st_bbox(camden), 5000, 5000)

ggplot() +
  base_map(bbox, increase_zoom = 2) +
  geom_sf(data = camden, fill = NA) +
  coord_sf(xlim = c(bbox['xmin'], bbox['xmax']),
           ylim = c(bbox['ymin'],bbox['ymax']),crs = 4326)

}
